requirements.txt
fastapi==0.65.1
uvicorn
ipython>=7.17.0
instagrapi>=1.16.12
python-multipart==0.0.5
Pillow==8.1.1
moviepy==1.0.3
pytest-asyncio==0.14.0
httpx==0.17.1
pudb==2021.1
ipdb==0.13.9
aiofiles==0.7.0
requests==2.26.0
starlette~=0.14.2
pytest~=6.2.4
tinydb==4.5.1
Flask-Cors==3.0.10
python-decouple==3.6
autopep8
click==8.1.2
colorama==0.4.4
Flask==2.1.1
importlib-metadata==4.11.3
python-dotenv
Flask-MySQLdb==0.2.0
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pycodestyle==2.7.0
toml==0.10.2
Werkzeug==2.0.1

extcolors==1.0.0
colormap==1.0.4
pandas
easydev
